requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
clickhouse-connect==0.9.2
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.104.1
frozenlist==1.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
jinja2==3.1.2
lz4==4.4.5
markupsafe==3.0.3
multidict==6.7.0
passlib==1.7.4
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pyyaml==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
telethon==1.34.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
aiohttp==3.9.1
aiosignal==1.4.0
attrs==25.4.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.0
rsa==4.9.1
telethon==1.34.0
typing-extensions==4.15.0
yarl==1.22.0
telethon==1.34.0
aiohttp==3.9.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
clickhouse-connect==0.7.0
passlib[bcrypt]==1.7.4
